BUG: Force emptyGrayLevels datatype to int
emptyGrayLevels is used as integer indices to remove the gray levels from the matrix. However, in linux python 3.6 and 3.7, the datatype defaults to int64, which is rejected by python when trying to use this as an index. Therefore, for the datatype of emptyGrayLevels to int, preventing the IndexError.

+++ radiomics/glszm.py
@@ -95,7 +95,7 @@ class RadiomicsGLSZM(base.RadiomicsFeaturesBase):
     # Delete rows that specify gray levels not present in the ROI
     NgVector = range(1, Ng + 1)  # All possible gray values
     GrayLevels = self.coefficients['grayLevels']  # Gray values present in ROI
-    emptyGrayLevels = numpy.array(list(set(NgVector) - set(GrayLevels)))  # Gray values NOT present in ROI
+    emptyGrayLevels = numpy.array(list(set(NgVector) - set(GrayLevels)), dtype=int)  # Gray values NOT present in ROI
 
     P_glszm = numpy.delete(P_glszm, emptyGrayLevels - 1, 1)
